requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==3.1.0
importlib-metadata==3.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
numpy==1.19.5
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.1.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scipy==1.5.4
six==1.15.0
tabulate==0.8.7
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
usienarl==0.7.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
